WIP: loplugin:redundantcast
(after a to-be-committed improved loplugin:cstylecast would have rewritten the
C-style cast into a static_cast)

+++ forms/source/component/formcontrolfont.cxx
@@ -92,7 +92,7 @@ namespace frm
                 break;
 
             case PROPERTY_ID_FONT_WEIGHT:
-                aValue <<= (float)_rFont.Weight;
+                aValue <<= _rFont.Weight;
                 break;
 
             case PROPERTY_ID_FONT_SLANT:
